refactor: merge level creation into one file
