+++ css-selectors/src/components/start-game.ts
@@ -1,7 +1,7 @@
 import { EventEmitter } from './classes/EventEmeitter';
 import { DataStorage } from './classes/Storage';
 import { loadProgress } from './game/load-game-progress';
-import { createLevel } from './game/tasks';
+import { createLevel } from './game/level-for-game';
 import { setFocus } from './view/set-focus';
 
 export function startGame(): void {
